requirements.txt
flask
flask_sqlalchemy
pyodbc
python-dotenv
psycopg2